nagios-addons: Add NRPE command for gluster processes
  Adding NRPE command for all process related services. We already
have passive checks through NSCA for all these services and bricks.
But passive checks are submitted only when there is a state change and
Nagios may miss those  check results if those services are not
cofigured when we submit passive checks. So, we should have active
checks running atleast for the first time to determine the initial
status of the service

 Command 'check_proc_status' helps to check the status of all
gluster related process
 Command 'check_brick_status' helps to check the status of bricks
in the host.

Change-Id: I1f442c0c1e54d606576bc0814044f2f149bca220
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>
Reviewed-on: http://review.gluster.org/7694
Reviewed-by: Kanagaraj M <kmayilsa@redhat.com>
Reviewed-by: Sahina Bose <sabose@redhat.com>