gluster-nagios-addons: Fix for sadf plugins going to unknown state.
This patch fixes the issue of sadf plugins going to unknown state
giving xml error. This issue occurs during the day's transition due
to sadf returning output without any records. This patch adds a job
to collect data two times during the last minute of the day,
because of which we get a record for time 00:00:01, which was not
got earlier.

Change-Id: Ibcecfe22215bdf4e41a5e9fed329a7a99320bedc
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1081953
Signed-off-by: ndarshan <dnarayan@redhat.com>
Reviewed-on: http://review.gluster.org/7802
Reviewed-by: Bala FA <barumuga@redhat.com>
Reviewed-by: Sahina Bose <sabose@redhat.com>