gluster-nagios-addons: Add thin lv details to check_disk_and_inode plugin.
check_disk_and_inode plugin is modified to show the
underlying thin pool details when invoked with -t
option. This is used to get the brick utilization.
If the brick is made up of thin lv then it returns
the virtual and actual utilization detail else
the plugin returns the normal output.

Change-Id: I2a2853a1b6e3e91c0178c3ed1cbc34de88ce3da2
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1143995
Signed-off-by: ndarshan <dnarayan@redhat.com>
Reviewed-on: http://review.gluster.org/9115
Reviewed-by: Sahina Bose <sabose@redhat.com>
Reviewed-by: Bala FA <barumuga@redhat.com>