nagios-plugins: Disk utilization plugin revamp.
This patch is aimed at revamping the existing check_disk_and_inode
plugin in such a way that the plugin behaviour is same for both
disk utilization service and brick utilization service. It has
just two options include and exclude to include or exclude the
mountpoints to be monitoried. This plugin monitors the data, inode,
thinpool, thinpool-metadata utilization irrespective of service
using this plugin.

Change-Id: I98b18d6f3b22915978f8c6381adc99729a4e1161
Bug-Url:https://bugzilla.redhat.com/show_bug.cgi?id=1167771
Signed-off-by: ndarshan <dnarayan@redhat.com>
Reviewed-on: http://review.gluster.org/9389
Reviewed-by: Bala FA <barumuga@redhat.com>
Reviewed-by: Sahina Bose <sabose@redhat.com>